requirements.txt
swarms
streamlit
swarm_models
google-genai
langchain_community